+++ reqwest-retry/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Added `with_retry_log_level` to `RetryTransientMiddleware`
+
 ## [0.5.0] - 2024-04-10
 
 ### Breaking changes